document-assistant/pyproject.toml
[project]
name = "document-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.13",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.1",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
]
